+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 6/23/2022 v 1.0.3
 
 - Fixed an issue causing verbose functionality to incorretly remove escaped backslashes.
+- Added .backslash method to expressions
 
 ## 6/21/2022 v 1.0.1